Fix climbing ledges with a small height.
Find_ceil would get the ceiling bellow instead of the ceiling above. Now it only tries to find the ceiling if it's strictly necessary (if it has the default max value of 100000 and m->ceil is null (It means it couldn't find one since the room above is not loaded.